feat!: get UiPath.CLI version based on platform version
Add input platform-version for specifying the versions of UiPath
Studio and/or Robot version used in major/minor version.

As an example, if the Studio version used in your environment is 2024.10.13 the input should be '2024.10'

Closes #171

BREAKING CHANGE: The default CLI version is now based on 'platform-version' input. 'version' input will no longer provide the default CLI version